Refactor: Update tests and config
